Maintainer field of Biocore Team packages now points to the BioC List
Since biocpkgs@lists.fhcrc.org seems to be a spam magnet and isn't
getting much, if any, use we are changing the maintainer email address
for BioC Core Team maintained packages to simply be the main
Bioconductor user email list.  Most users report problems there anyhow
and we feel that encouraging that single channel for bug reports is
best.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/gpls@26527 bc3139a8-67e5-0310-9ffc-ced21a209358